requirements.txt
ipython==8.12.3
langchain_core==0.3.7
langchain_openai==0.2.1
langgraph==0.2.32
typing_extensions==4.12.2
google-api-python-client==2.165.0
python-dotenv==1.1.0
pandas==2.2.3
